requirements.txt
psycopg2_binary
xlsxwriter
requests
pandas
